+++ ConcordiumWallet/Views/Identities/IdentityList/IdentitiesViewController.swift
@@ -144,7 +144,7 @@ extension IdentitiesViewController: IdentitiesViewProtocol {
         
         ac.addAction(continueAction)
         
-        if canSendMail {
+        if MailHelper.canSendMail {
             let supportAction = UIAlertAction(title: "identityfailed.contactsupport".localized, style: .default) { [weak self] _ in
                 guard let self = self else { return }
                 self.launchSupport(
